Change download button behaviour (#119)
Change how the preferred download source for posters, publications or talks is
automatically chosen. For posters and talks they will be downloaded from the DOI, unless
a PDF file has been uploaded to the website. For publications, they will be downloaded
from the article DOI (in case it has open access) or from the preprint DOI (if it
doesn't). Also, if a PDF version of the article is uploaded, then it will be the
preferred choice. Upload instructions on the README.md. Remove PDF version of lapis2019
poster and pdfs directory.