Improve about page
Add more social links, now with handle and optional tags for including
them on the home page or not. Also include a contact tag for those links
that should be inside the "Write me" section in about page.
Add my GPG public key to the website.
Make some changes to the CSS style.
Add a new about.html layout.
Include an optional title argument for the link macro.