Improve about page (#148)
Add more social links, now with optional handle and tags for including them on
the home page or not. Also include a contact tag for those links that should be
inside the "Write me" section in about page. Add my GPG public key to the
website. Make some changes to the CSS style. Add a new about.html layout.
Include an optional title argument for the link macro. Add a new social_item
macro. Add forkawesome and use it for show the matrix.org logo. Small
improvements to the footer and ignore every Markdown file inside the css
folder.